# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API doc MDM Customer",
        "description": "API doc MDM Customer",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "mario.fredrick@mncgroup.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/{business_unit}/customers/profile/search": {
            "post": {
                "tags": [
                    "Retrieve Business Unit Customer Profile"
                ],
                "summary": "Retrieve Customer Data",
                "description": "Use json payload to retrieve customer data",
                "operationId": "retrieveProfile",
                "parameters": [
                    {
                        "name": "SecretKey",
                        "in": "header",
                        "description": "client api token key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_id": {
                                        "description": "customer id",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
...
}